requirements.txt
fastapi>=0.100.0
openai>=1.0.0
starlette>=0.27.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
